refactor: :recycle: melhorias levantadas pelo CR
Melhorias levantadas pelo Code Review. Englobando versao do node 20.16.0, esnext, nodemon e padronizacao do nome dos arquivos (hello-schema)
