+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 package-lock.json
 .vscode
 .env
+dist
+tsconfig.tsbuildinfo
