+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.20.4
+v20.16.0
