+++ package.json
@@ -2,13 +2,15 @@
   "name": "user-server",
   "version": "1.0.0",
   "description": "Este projeto é um sistema de gerenciamento de usuários",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "type": "module",
   "scripts": {
-    "start": "ts-node src/index.ts",
-    "dev": "ts-node-dev --respawn src/index.ts"
+    "build": "tsc",
+    "start": "node dist/index.js",
+    "dev": "nodemon"
   },
   "engines": {
-    "node": "18"
+    "node": "20"
   },
   "repository": {
     "type": "git",
@@ -28,8 +30,8 @@
   },
   "devDependencies": {
     "@types/node": "22.1.0",
+    "nodemon": "3.1.4",
     "ts-node": "10.9.2",
-    "ts-node-dev": "2.0.0",
     "typescript": "5.5.4"
   }
 }
